# contract/openapi.yaml
openapi: 3.1.0
info:
  title: cmProjectX Service API
  version: 0.1.0
  description: >
    Contract between the Rust/WinUI 3 client and the .NET Intune engine.
    This is the single source of truth for DTOs and endpoints — generate
    C# and Rust types from it (see contract/README.md).
servers:
  - url: http://127.0.0.1:5099
    description: Local sidecar (MVP)
paths:
...
  /gitops/plan:
    post:
      operationId: gitopsPlan
      summary: Diff a repo tree against the live tenant (read-only change set)
      description: >
        Re-pulls the live tenant through the same normalize pipeline and diffs it
        against the committed repo files with JsonDrift. Objects in the repo but not
        the tenant are Added; in the tenant but not the repo are Removed; in both,
        field-level Modified. The returned planId pins the live state the plan was
        computed against, so apply can reject on drift. No writes. 409 when signed out.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GitOpsPlanRequest" }
      responses:
        "200":
          description: The change set the repo would apply to the tenant
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GitOpsPlan" }
        "400": { description: Missing repoPath or no manifest.json }
        "409": { description: Not signed in }
  /gitops/apply:
    post:
      operationId: gitopsApply
      summary: Apply a reviewed plan to the tenant (gated; per-object conflict-safe)
      description: >
        Requires confirm:true and the planId returned by plan. Re-checks each
        object's live state against the plan (hash mismatch → that object is reported
        conflict and nothing is written for it). Added → create, Modified → update,
        through the Core engines + ImportService, each snapshot-on-write. Removed
        objects are skipped unless confirmDestroy:true. Conditional Access is never
        written (locked guarantee).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GitOpsApplyRequest" }
      responses:
        "200":
          description: Per-object outcomes (applied / skipped / conflict)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GitOpsApplyResult" }
        "400": { description: confirm not set, or missing repoPath/planId }
        "404": { description: Unknown or expired planId }
        "409": { description: Not signed in }
...
  /policies/compare:
    post:
      operationId: comparePolicies
      summary: Compare two live SettingsCatalog policies (A vs B), setting-by-setting
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PolicyCompareRequest" }
      responses:
        "200":
          description: Settings grouped by verdict (matching/same · missing/only-A · drifted/different · extra/only-B)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BaselineComparison" }
        "400": { description: aId and bId required, or policy settings unreadable }
        "409": { description: Not signed in }

  # ── M12.1 cache-dev (Cache Sync): inspect / warm / evict the LiteDB read-through
  #    blob cache for the active tenant. All ops 409 when signed out. ──
...
  /packs/{id}:
    get:
      operationId: getPack
      summary: One pack's manifest + repo summary
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Pack manifest
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackManifest" }
        "404": { description: No such pack }
  /packs/{id}/adopt:
    post:
      operationId: adoptPack
      summary: >
        Bind parameters → M15 gitops plan (JsonDrift) → gated apply. Read-only by
        default (plan-only); pass confirm:true with a reviewed planId to apply.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PackAdoptRequest" }
      responses:
        "200":
          description: Adoption result (the M15 plan, or the apply outcome)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PackAdoptResult" }
        "404": { description: No such pack }
        "409": { description: Not signed in }
        "501": { description: GitOps (M15) endpoints not present in this build }
...
components:
...
  schemas:
...
    DriftChange:
      type: object
      properties:
        path:     { type: string, description: "JSON path of the changed field" }
        kind:     { type: string, enum: [Added, Removed, Modified] }
        before:   { nullable: true }
        after:    { nullable: true }
      required: [path, kind]

    # ─── M15 Policy-as-Code / GitOps ──────────────────────────────────────────
...
    GitOpsPlanRequest:
      type: object
      properties:
        repoPath: { type: string }
        surfaces:
          type: array
          items: { type: string }
      required: [repoPath]
    GitOpsPlan:
      type: object
      description: A Terraform-style change set — per-object DriftChange[] grouped by verdict.
      properties:
        planId:       { type: string }
        repoPath:     { type: string }
        generatedUtc: { type: string, format: date-time }
        summary:      { $ref: "#/components/schemas/GitOpsPlanSummary" }
        objects:
          type: array
          items: { $ref: "#/components/schemas/GitOpsPlanObject" }
      required: [planId, repoPath, generatedUtc, summary, objects]
    GitOpsPlanSummary:
      type: object
      properties:
        add:     { type: integer, format: int32 }
        change:  { type: integer, format: int32 }
        destroy: { type: integer, format: int32 }
        noop:    { type: integer, format: int32 }
      required: [add, change, destroy, noop]
    GitOpsPlanObject:
      type: object
      properties:
        surface:    { type: string, description: "Surface key, e.g. device-configs" }
        objectType: { type: string }
        objectName: { type: string }
        verdict:    { type: string, enum: [Added, Removed, Modified] }
        changes:
          type: array
          items: { $ref: "#/components/schemas/DriftChange" }
      required: [surface, objectType, objectName, verdict, changes]
    GitOpsApplyRequest:
      type: object
      properties:
        repoPath:       { type: string }
        planId:         { type: string }
        confirm:        { type: boolean, description: "Must be true (M6 gate)" }
        confirmDestroy: { type: boolean, description: "Allow deleting tenant objects absent from the repo (default false)" }
      required: [repoPath, planId, confirm]
    GitOpsApplyResult:
      type: object
      properties:
        planId:     { type: string }
        appliedUtc: { type: string, format: date-time }
        objects:
          type: array
          items: { $ref: "#/components/schemas/GitOpsApplyObject" }
        summary:    { $ref: "#/components/schemas/GitOpsApplySummary" }
      required: [planId, appliedUtc, objects, summary]
    GitOpsApplyObject:
      type: object
      properties:
        objectName: { type: string }
        outcome:    { type: string, enum: [applied, skipped, conflict] }
        snapshotId: { type: string, nullable: true }
        newId:      { type: string, nullable: true }
        reason:     { type: string, nullable: true }
      required: [objectName, outcome]
    GitOpsApplySummary:
      type: object
      properties:
        applied:  { type: integer, format: int32 }
        skipped:  { type: integer, format: int32 }
        conflict: { type: integer, format: int32 }
      required: [applied, skipped, conflict]
...
    PolicyCompareRequest:
      type: object
      description: POST /policies/compare body — two live SettingsCatalog policy ids (A vs B).
      properties:
        aId: { type: string }
        bId: { type: string }
      required: [aId, bId]
    BaselineSettingComparison:
      type: object
      description: >-
        One setting's A-vs-B values (both nullable — missing on the B side, or extra beyond A).
        settingName is the human-readable label from the embedded Settings Catalog registry.
      properties:
        settingDefinitionId: { type: string }
        settingName:         { type: string, nullable: true }
        baselineValue:       { type: string, nullable: true }
        tenantValue:         { type: string, nullable: true }
      required: [settingDefinitionId]
    BaselineComparison:
      type: object
      description: >
        A SettingsCatalog baseline vs a tenant policy, grouped by verdict (mirrors
        api_types::BaselineComparison / BaselineComparisonDto byte-for-byte).
      properties:
        baselineName:     { type: string }
        tenantPolicyId:   { type: string, nullable: true }
        tenantPolicyName: { type: string, nullable: true }
        matching: { type: array, items: { $ref: "#/components/schemas/BaselineSettingComparison" } }
        missing:  { type: array, items: { $ref: "#/components/schemas/BaselineSettingComparison" } }
        drifted:  { type: array, items: { $ref: "#/components/schemas/BaselineSettingComparison" } }
        extra:    { type: array, items: { $ref: "#/components/schemas/BaselineSettingComparison" } }
      required: [baselineName, matching, missing, drifted, extra]
...
    PackManifest:
      type: object
      description: >
        A pack is a versioned M15 desired-state repo plus this top-level pack.json
        manifest. targetSurfaces mirror the path catalog (Surfaces.cs); the repo body
        is byte-identical to a `cmpx pull` export. Adopting = the M15 round-trip
        (plan → gate → apply) against the operator's tenant, parameter-substituted.
      properties:
        schemaVersion:  { type: string }
        id:             { type: string }
        name:           { type: string }
        version:        { type: string }
        publisher:      { type: string, nullable: true }
        description:    { type: string, nullable: true }
        targetSurfaces:
          type: array
          items: { type: string }
        minSidecar:     { type: string, nullable: true }
        parameters:
          type: array
          items: { $ref: "#/components/schemas/PackParameter" }
        repoPath:       { type: string, nullable: true, description: "Local working-tree path the manifest was loaded from" }
        objectCount:    { type: integer, format: int32, minimum: 0 }
        signature:      { type: string, nullable: true, description: "Reserved — detached signature for supply-chain trust" }
      required: [schemaVersion, id, name, version, targetSurfaces, parameters, objectCount]
    PackParameter:
      type: object
      description: One adoption parameter the operator binds before plan (e.g. a target group id).
      properties:
        name:      { type: string }
        type:      { type: string, description: "groupId | filterId | int | string | bool" }
        prompt:    { type: string, nullable: true }
        required:  { type: boolean }
        default:   { type: string, nullable: true }
        appliesTo: { type: string, nullable: true, description: "JSON path the value substitutes into" }
      required: [name, type, required]
    PackAdoptRequest:
      type: object
      description: >
        Bind operator-supplied parameters, then plan (default) or apply. confirm:false
        (default) returns the M15 plan only — a Terraform-style preview. confirm:true
        with the reviewed planId runs the gated M15 apply.
      properties:
        parameters:
          type: object
          additionalProperties: { type: string }
          description: name → value bindings for the manifest parameters
        confirm:        { type: boolean, description: "false (default) = plan only; true = apply the reviewed plan" }
        planId:         { type: string, nullable: true, description: "Required when confirm:true — the planId returned by a prior plan" }
        confirmDestroy: { type: boolean, description: "Allow deletes when applying (default false)" }
    PackAdoptResult:
      type: object
      description: >
        The outcome of an adopt. phase = plan when confirm:false (plan carries the M15
        plan id + change summary the operator reviews), apply when confirm:true (applyResult
        carries the per-object outcome). requiresM15 is true when the GitOps endpoints are
        absent in this build (this branch bases off main, where M15 is not yet merged).
      properties:
        packId:       { type: string }
        phase:        { type: string, enum: [plan, apply, error] }
        requiresM15:  { type: boolean }
        message:      { type: string, nullable: true }
        planId:       { type: string, nullable: true }
        summary:      { type: string, nullable: true, description: "Human-readable change summary from the M15 plan" }
        plan:         { nullable: true, description: "The raw M15 plan body (passthrough), when phase=plan" }
        applyResult:  { nullable: true, description: "The raw M15 apply result (passthrough), when phase=apply" }
      required: [packId, phase, requiresM15]
...